Разбивает файл на модули
